koka: take C compiler from targetPackages
buildPackages.stdenv.cc.cc is a C compiler that runs on the build
platform and produces binaries for the host platform. This is not
what we want. Also pkgsHostTarget.stdenv.cc is not the compiler we
want as stdenv always runs on the previous stage so to say (the stdenv
is used to build the package set, in the case of cross compiling
this is not done natively). Thus stdenv.targetPackages.stdenv.cc is what
we want (note that I haven't really spent any thought on cross compiling
a koka cross compiler, but it'll be a while until someone needs to do
this).

Testing would be appreciated.